Fix name

+++ Package.swift
@@ -2,7 +2,7 @@
 import PackageDescription
 
 let package = Package(
-    name: "BigInt",
+    name: "BigInteger",
     dependencies: [],
 	
     targets: [
